+++ test/LocalStoreByTimestampTest.ts
@@ -5,8 +5,9 @@ import LocalStoreByTimestamp = require("../local-store/LocalStoreByTimestamp");
 import UniqueChronologicalKeys = require("../local-store/UniqueChronologicalKeys");
 import MemoryStore = require("../local-store/MemoryStore");
 import LocalStorageStore = require("../local-store/LocalStorageStore");
-import CommonStorageTests = require("./CommonStorageTests");
 
+declare var global: { process: { hrtime(): [number, number]; } };
+declare var console: { log(...args: any[]): void };
 
 var asr = chai.assert;
 
@@ -17,7 +18,7 @@ suite("LocalStoreByTimestamp", function LocalStoreByTimestampTest() {
         var memStore = MemoryStore.newInst();
         var localStore = LocalStorageStore.newInst(memStore, null, <LocalStore.FullStoreHandler><any>null, false, false, 20);
         var store = LocalStoreByTimestamp.newTimestampInst(localStore);
-        store.keyGenerator = UniqueChronologicalKeys.uniqueTimestampNodeJs;
+        store.keyGenerator = UniqueChronologicalKeys.uniqueTimestamp;
 
         var keyA = store.addItem({ a: 1 });
         var keyB = store.addItem({ b: 1 });
@@ -32,6 +33,42 @@ suite("LocalStoreByTimestamp", function LocalStoreByTimestampTest() {
 
         asr.deepEqual(store.getKeys(), [keyB, keyC, keyA]);
         asr.deepEqual(store.getItem(keyB), { b: 1 });
+    });
+
+
+    test("check-unique-chronological-keys-(some-duplicates)", function checkUniqueChronologicalKeysUniqueness() {
+        var n = 0, gt = 0, r = 0;
+        var keyGen = function () {
+            return <number>(n = Date.now()) + (r = Math.random()); // (gt = global.process.hrtime()[1] / 1000)
+        }; //UniqueChronologicalKeys.uniqueTimestamp;
+
+        var outerTimestamps = [];
+        var duplicates = [];
+        var mostPerGroup = 0;
+        var cnt = 0;
+        for (var i = 0; i < 1000; i++) {
+            var ary: { key: number; calc: string }[] = [];
+            var dupCount = duplicates.length;
+            for (var k = 0; k < 100; k++) {
+                var key = keyGen();
+                // check for duplicates
+                var idx = ary.findIndex(s => s.key === key);
+                if (idx > -1) {
+                    duplicates.push("duplicate at loop (" + i + "," + k + ")" + "\n\t| curr: key=" + key + ", n=" + n + ", gt=" + gt + ", r*=" + r + "\n\t| prev: key=" + ary[idx].key + ary[idx].calc);
+                }
+                ary[k] = { key: key, calc: ", n=" + n + ", gt=" + gt + ", r*=" + r };
+            }
+            var idx = Math.floor(Math.random() * 100);
+            cnt += (i % 2 === 0 ? ary[idx].key : -ary[idx].key);
+            mostPerGroup = Math.max(duplicates.length - dupCount, mostPerGroup);
+
+            var now = Date.now();
+            if (outerTimestamps.indexOf(now) === -1) {
+                outerTimestamps.push(now);
+            }
+        }
+        console.log("timestamps:", outerTimestamps.length, "duplicates:", duplicates.length, "most-per-group:", mostPerGroup);
+        return cnt;
     });
 
 });